Base64 encode/decode strings Marshaled strings
There is a nasty bug I tripped over because Marshal, contrary to popular
belief, does not emit only printable characters. In fact, some of the
sequences it emits are *totally invalid as UTF-8* (UTF-8 does not
magically take care of converting a stream of bytes).

This is essentially what the code *was* doing:

>> marshal_str = Marshal.dump(:foo => 900)
=> "\004\b{\006:\bfooi\002\204\003"
>> marshal_bytes = marshal_str.to_java_bytes
=> #<#<Class:01x525dc268>:0x9866417>
>> java_str = java.lang.String.new(marshal_bytes, 'UTF-8')
=> #<Java::JavaLang::String:0x74d175ff>
>> unpacked_marshal_str = String.from_java_bytes(java_str.getBytes('UTF-8'))
=> "\004\b{\006:\bfooi\002\357\277\275\003"
>> marshal_str == unpacked_marshal_str
=> false
>> Marshal.load(unpacked_marshal_str)
=> {:foo=>49135}

Version 1.5.1 of the library produces this exact result:

>> gem 'ikai-jruby-memcache-client', '1.5.1'
=> true
>> require 'memcache'
=> true
>> mc = MemCache.new(%w[localhost:11211]); nil
=> nil
>> mc.set('fargo', {:foo => 900})
=> true
>> mc.get('fargo')
=> {:foo=>49135}

Interestingly enough, on my Solaris system, using 64 bit java, going
through the MemCachedClient corrupts the data even further, preventing
marshal from loading it at all.

If you take the intermediate step of Base64 encoding the marshalled
string, thereby ensuring that all characters are printable, conversion
to/from UTF-8 is much more likely to succeed:

>> marshal_str = Marshal.dump(:foo => 900)
=> "\004\b{\006:\bfooi\002\204\003"
>> b64_marshal_str = Base64.encode64(marshal_str)
=> "BAh7BjoIZm9vaQKEAw==\n"
>> b64_marshal_bytes = b64_marshal_str.to_java_bytes
=> #<#<Class:01x525dc268>:0x40930120>
>> java_str = java.lang.String.new(b64_marshal_bytes, 'UTF-8')
=> #<Java::JavaLang::String:0x521d590d>
>> unpacked_b64_marshal_str = String.from_java_bytes(java_str.getBytes('UTF-8'))
=> "BAh7BjoIZm9vaQKEAw==\n"
>> unb64_marshal_str = Base64.decode64(unpacked_b64_marshal_str)
=> "\004\b{\006:\bfooi\002\204\003"
>> Marshal.load(unb64_marshal_str)
=> {:foo=>900}

And this is how the client now behaves:

>> mc = MemCache.new(%w[localhost:11211]); nil
=> nil
>> mc.set('fargo', {:foo => 900})
=> true
>> mc.get('fargo')
=> {:foo=>900}